pyproject.toml
[project]
name = "t-trade"
version = "0.1.0"
description = "IBKR 日内交易提示系统 - 第一阶段"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ib-insync>=0.9.86",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pendulum>=3.0.0",
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
tbot = "tbot.main:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
